+++ lib/pocketbase.dart
@@ -19,6 +19,7 @@ export "src/dtos/auth_method_provider.dart";
 export "src/dtos/auth_methods_list.dart";
 export "src/dtos/backup_file_info.dart";
 export "src/dtos/batch_result.dart";
+export "src/dtos/collection_field.dart";
 export "src/dtos/collection_model.dart";
 export "src/dtos/email_template_config.dart";
 export "src/dtos/health_check.dart";
@@ -34,7 +35,6 @@ export "src/dtos/record_auth.dart";
 export "src/dtos/record_model.dart";
 export "src/dtos/record_subscription_event.dart";
 export "src/dtos/result_list.dart";
-export "src/dtos/schema_field.dart";
 export "src/dtos/token_config.dart";
 
 // services (exported mainly for dartdoc - https://github.com/dart-lang/dartdoc/issues/2154)